Cloud_pipelines/Azure/facedetect-pipeline/requirements.txt
numpy
Pillow
dlib
azure-storage-blob==1.5.0
